+++ spec/models/user_spec.rb
@@ -981,6 +981,17 @@ describe User, :type => :model do
     end
   end
 
+  describe "#export" do
+    it "doesn't change the filename when the user is saved" do
+      user = FactoryGirl.create(:user)
+
+      filename = user.export.filename
+      user.save!
+
+      expect(User.find(user.id).export.filename).to eq(filename)
+    end
+  end
+
   describe "queue_export" do
     it "queues up a job to perform the export" do
       user = FactoryGirl.create(:user)